+++ android-app/app/src/main/java/com/nimtome/app/ui/components/CharacterCard.kt
@@ -1,4 +1,4 @@
-package com.flyinpancake.dndspells.ui.components
+package com.nimtome.app.ui.components
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
@@ -10,10 +10,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.flyinpancake.dndspells.model.DndCharacter
-import com.flyinpancake.dndspells.sampleCharacter
-import com.flyinpancake.dndspells.ui.theme.DndSpellsTheme
-import com.flyinpancake.dndspells.ui.theme.CardElevation
+import com.nimtome.app.model.DndCharacter
+import com.nimtome.app.sampleCharacter
+import com.nimtome.app.ui.theme.DndSpellsTheme
+import com.nimtome.app.ui.theme.CardElevation
 
 
 @ExperimentalMaterialApi
